+++ cypress/e2e/ux_sins_admin.js
@@ -18,7 +18,9 @@ context( 'Admin page (/admin)', () => {
         cy.contains( 'Password requirements are not shown' ).should( 'be.visible' )
         cy.contains( 'Toast notifications in center of screen' ).should( 'be.visible' )
         cy.contains( 'blocking content' ).should( 'be.visible' )
-        cy.get( 'input[type="checkbox"]' ).should( 'have.length', 2 )
+        cy.contains( 'Forced small hamburger menu' ).should( 'be.visible' )
+        cy.contains( 'tiny hamburger icon' ).should( 'be.visible' )
+        cy.get( 'input[type="checkbox"]' ).should( 'have.length', 3 )
     } )
 
     it( 'Toggles start unchecked by default', () => {